Refactor: add renvm-address-set endpoint folder structure
